[#63] refactor(brand) : redis 시간복잡도에 따른 호출 메서드 변경
